+++ apps/explorer/lib/explorer/chain/import/stage/block_referencing.ex
@@ -18,7 +18,14 @@ defmodule Explorer.Chain.Import.Stage.BlockReferencing do
       Runner.Tokens,
       Runner.TokenTransfers,
       Runner.Address.TokenBalances,
+      Runner.OptimismOutputRoots,
       Runner.TransactionActions,
+      Runner.OptimismFrameSequences,
+      Runner.OptimismTxnBatches,
+      Runner.OptimismOutputRoots,
+      Runner.OptimismDeposits,
+      Runner.OptimismWithdrawals,
+      Runner.OptimismWithdrawalEvents,
       Runner.Withdrawals
     ]